feat(tid): behörighetsnycklar för tid och lön
Fas 4.0 av flytten av tidrapporteringen från Blikk. Fem nya nycklar:
time.entry.write (member, sales, admin — alla anställda rapporterar sin egen tid),
och time.entry.read.all / time.approve / time.payroll.read / time.reference.manage
på admin. konsult får ingenting: extern, och tid är personuppgift.

Nycklarna ligger MEDVETET utanför crm.*-rymden. Tid är hela företagets, inte en
CRM-yta — en tidnyckel under crm.* hade ärvt metavakterna crm.access/crm.write och
tyst stängt ute installatörerna från sina egna timmar. Ett test vaktar gränsen.

SQL:en måste köras före koden. getEffectivePermissions failar closed, så en nyckel
som saknas i databasen nekar alla — det är en säkerhetsegenskap, inte en frist.

requirePermission/requireSignedInUser flyttas ur app/api/crm/_shared.ts så
tid-routerna kan grinda utan att importera ur CRM-mappen. De hamnar i en EGEN
modul, inte i lib/auth/route.ts bredvid getCurrentUser: guarden måste anropa
getCurrentUser över en modulgräns, för route-testerna mockar '@/lib/auth/route'
för att simulera ingen session / member / sales. vi.mock ersätter en moduls
exporter men kan inte fånga ett anrop en modul gör till sig själv — slås filerna
ihop slutar session-mocken gälla och 65 vakttester börjar tyst asserta mot ett
riktigt uppslag i stället för scenariot de heter efter. Det upptäcktes genom att
göra det felet först.
